review: move library modules to lib/
