+++ packages/workflows/src/review/index.ts
@@ -1,3 +1,3 @@
-export * from "./retrieveContext.js";
-export * from "./reviewWorkflow.js";
-export * from "./validateReviewOutput.js";
+export * from "./lib/reviewContext.js";
+export * from "./lib/reviewWorkflow.js";
+export * from "./lib/validateReviewOutput.js";
