+++ packages/workflows/src/review/lib/normalizeSignal.ts
